Workaround 64bit Linux quirks
64bit Linux sometimes returns a higher value for the CMSG_LEN header
from recvmsg(2) than the original value that was sent in sendmsg (2).
This breaks libancillary as the CMSG_LEN no longer represents the
number of file descriptors passed from the client.

This patch sends the number of file descriptors through explicitly in
the message body. This removes the need for the CMSG_LEN math.
Previously the message body was being sent with a junk value that was
ignored on the receiving side.